Add low level support for BEP 10 user protocols

+++ peer_protocol/extended.go
@@ -24,7 +24,7 @@ type (
 	}
 
 	ExtensionName   string
-	ExtensionNumber int
+	ExtensionNumber uint8
 )
 
 const (
